+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   You can use `Ctrl` + `Enter` to unfocus and refocus a Task when editing
 -   When a new Task is added, that Task is automatically focused and editable
+-   History button toggles the history panel instead of just opening it
 
 ## 1.5.0